# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Replayjs example API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /api
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

paths:

  /blob/{id}:
    x-swagger-router-controller: blob
    get:
      description: get json blob by id
      operationId: showBlob
      parameters:
        - name: id
          in: path
          description: blob id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Blob"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /blob:
    x-swagger-router-controller: blob
    post:
      description: Post a json blob
      operationId: createBlob
      parameters:
        - name: body
          in: body
          description: The json blob to post
          required: true
          schema:
            $ref: "#/definitions/Blob"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/BlobResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /swagger:
    x-swagger-pipe: swagger_raw

# complex objects have schema definitions
definitions:
  BlobResponse:
    description: Response after posting a Blob
    required:
      - id
    properties:
      id:
        type: string

  PatchRequest:
    type: array
    items:
      $ref: "#/definitions/PatchDocument"

  PatchDocument:
    description: A JSONPatch document as defined by RFC 6902
    required:
     - "op"
     - "path"
    properties:
     op:
      type: string
      description: The operation to be performed
      enum:
       - "add"
       - "remove"
       - "replace"
       - "move"
       - "copy"
       - "test"
     path:
      type: string
      description: A JSON-Pointer
     value:
      type: object
      description: The value to be used within the operations.
     from:
      type: string
      description: A string containing a JSON Pointer value.

# Post definitions
  Blob:
    title: A json blob
    type: object

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string

  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string